#151 #154
Fix reported issues
